Fixed windows-specific 'FileExistsError' exception that would occur if the file './rustdesk-{version}-install.exe' already existed. Swapped windows system calls for python to use the default 'python' instead of 'python3'.
FileExistsError exception would occur during the rename operation if the file existed. So, to avoid using try\except blocks with unlinking or deleting, we can just check the existance of the path, and respect filesystem ACLs.

+++ build.py
@@ -445,7 +445,7 @@ def build_flutter_windows(version, features, skip_portable_pack):
     os.chdir('libs/portable')
     system2('pip3 install -r requirements.txt')
     system2(
-        f'python3 ./generate.py -f ../../{flutter_build_dir_2} -o ./ -e ../../{flutter_build_dir_2}rustdesk.exe')
+        f'python ./generate.py -f ../../{flutter_build_dir_2} -o ./ -e ../../{flutter_build_dir_2}rustdesk.exe')
     os.chdir('../..')
     if os.path.exists('./rustdesk_portable.exe'):
         os.replace('./target/release/rustdesk-portable-packer.exe',
@@ -455,7 +455,12 @@ def build_flutter_windows(version, features, skip_portable_pack):
                   './rustdesk_portable.exe')
     print(
         f'output location: {os.path.abspath(os.curdir)}/rustdesk_portable.exe')
-    os.rename('./rustdesk_portable.exe', f'./rustdesk-{version}-install.exe')
+    
+    #respect system ACLs and prevent 'except FileExistsError' handling
+    if os.path.exists(f'./rustdesk-{version}-install.exe'):
+        os.replace('./rustdesk_portable.exe', f'./rustdesk-{version}-install.exe')
+    else:
+        os.rename('./rustdesk_portable.exe', f'./rustdesk-{version}-install.exe')
     print(
         f'output location: {os.path.abspath(os.curdir)}/rustdesk-{version}-install.exe')
 
@@ -473,7 +478,11 @@ def main():
     features = ','.join(get_features(args))
     flutter = args.flutter
     if not flutter:
-        system2('python3 res/inline-sciter.py')
+        #By default system uses python instead of python3 in windows, and *NIX is the opposite (unless python-is-python3 package is installed)
+        if not windows:
+            system2('python3 res/inline-sciter.py')
+        else:
+            system2('python res/inline-sciter.py')
     print(args.skip_cargo)
     if args.skip_cargo:
         skip_cargo = True
@@ -509,7 +518,7 @@ def main():
         os.chdir('libs/portable')
         system2('pip3 install -r requirements.txt')
         system2(
-            f'python3 ./generate.py -f ../../{res_dir} -o . -e ../../{res_dir}/rustdesk-{version}-win7-install.exe')
+            f'python ./generate.py -f ../../{res_dir} -o . -e ../../{res_dir}/rustdesk-{version}-win7-install.exe')
         system2('mv ../../{res_dir}/rustdesk-{version}-win7-install.exe ../..')
     elif os.path.isfile('/usr/bin/pacman'):
         # pacman -S -needed base-devel
